renamed test-api-service to exercise-api-service and added new endpoint to get all exercises
