+++ api-gateway/src/main/resources/application.properties
@@ -8,5 +8,5 @@ eureka.client.fetchRegistry=true
 
 zuul.ignored-services=*
 
-zuul.routes.test-api-service.service-id=testapi
-zuul.routes.test-api-service.path=/test/**
+zuul.routes.exercise-api-service.service-id=exerciseapi
+zuul.routes.exercise-api-service.path=/exercise/**
