+++ exercise-api-service/src/main/java/com/fitness/exerciseapiservice/service/ExerciseList.java
@@ -1,6 +1,6 @@
-package com.fitness.testapiservice.service;
+package com.fitness.exerciseapiservice.service;
 
-import com.fitness.testapiservice.repository.ExerciseRepository;
+import com.fitness.exerciseapiservice.repository.ExerciseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
