+++ exercise-api-service/src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=testapi
+spring.application.name=exerciseapi
 server.port=8081
 eureka.client.service-url.default-zone=http://localhost:8761/eureka
