--- test-api-service/src/main/java/com/fitness/testapiservice/controller/TestApiController.java
@@ -1,13 +1,0 @@
-package com.fitness.testapiservice.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestApiController {
-
-    @GetMapping("/hello")
-    public String testResult(){
-        return "Hello";
-    }
-}
